% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibri_test.R
\name{calibri_test}
\alias{calibri_test}
\title{Test for Calibri import and registration}
\usage{
calibri_test()
}
\description{
\code{calibri_test} tests to see if Calibri is imported and registered. Calibri is the Reason Foundation's main font.
}
\details{
Import and register Calibri in R with calibri_install().
}
