% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_reason.R
\docType{data}
\name{palette_reason}
\alias{palette_reason}
\title{The Reason Foundation [ggplot2] theme}
\format{A \code{list}.}
\usage{
palette_reason
}
\description{
Data used by the palettes in the reasontheme package.
}
\keyword{datasets}
