% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{reason_color_pal}
\alias{reason_color_pal}
\title{The Reason Foundation [ggplot2] theme}
\usage{
reason_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used by the Reason Foundation.
}
