% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reason_style.R
\name{reasonStyle}
\alias{reasonStyle}
\title{Add Reason theme to a ggplot chart}
\usage{
reasonStyle()
}
\description{
This function allows you to add the Reason theme to your ggplot graphics.
}
\examples{
\dontrun{
line <- library(ggplot2)
ggplot2(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
reasonStyle()
}
}
