% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_reason_theme.R
\name{set_reason_theme}
\alias{set_reason_theme}
\title{The Reason Foundation \link{ggplot2} theme}
\usage{
set_reason_theme(style = "slide")
}
\arguments{
\item{style}{The default theme style. "slide", "web", or "print".}
}
\description{
\code{set_reason_theme} provides a \link{ggplot2} theme formatted according to the
Reason Foundation style guide.
}
