% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reason_slide.R
\name{theme_reason_slide}
\alias{theme_reason_slide}
\title{A \link{ggplot2} theme formatted in the Reason Foundation style}
\usage{
theme_reason_slide(base_size = 14, base_family = "Calibri",
  base_line_size = 0.5, base_rect_size = 0.5)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_reason} provides a \link{ggplot2} theme formatted according to the
Reason Foundation style guide for presentation slides.
}
